Rename Arguments to MessageHelpers
