+++ src/BapesBot.Service/MessageHelpers/MessageHelper.cs
@@ -1,16 +1,14 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
 
-namespace BapesBot.Service.Arguments
+namespace BapesBot.Service.MessageHelpers
 {
-    public static class ArgumentParser
+    public static class MessageHelper
     {
         public static object[] GetArguments(string message)
         {
             // No arguments
             if (!message.Contains(' ')) return new object[] { };
-            
+
             message = message.Substring(message.IndexOf(' ') + 1);
 
             return message.Split(' ', StringSplitOptions.RemoveEmptyEntries);
